app/requirements.txt
SQLAlchemy
python-decouple
fastapi
uvicorn
passlib
pydantic
numpy
joblib
pydantic-settings
psycopg2-binary
asyncpg
scikit-learn
pika
python-dotenv